+++ src/common-components/icons/index.ts
@@ -1,2 +1,3 @@
 export { UndoIcon } from './UndoIcon';
 export { RedoIcon } from './RedoIcon';
+export { GripIcon } from './GripIcon';